Distance is now a long

+++ IEnumerableCorrelater/CorrelaterResult.cs
@@ -2,14 +2,14 @@
 {
     public class CorrelaterResult<T>
     {
-        public CorrelaterResult(int distance, T[] bestMatch1, T[] bestMatch2)
+        public CorrelaterResult(long distance, T[] bestMatch1, T[] bestMatch2)
         {
             Distance = distance;
             BestMatch1 = bestMatch1;
             BestMatch2 = bestMatch2;
         }
 
-        public int Distance { get; }
+        public long Distance { get; }
 
         public T[] BestMatch1 { get; }
